ENH: separate formatting from API endpoints (#22)
* Start working on formatter

* Fix bugs when local refresh token expired

When local token has invalid refresh token, generate a new token to
replace the old token.

* Add works on formatter and db handler

- Add formatter code for region market history endpoint
- Add starter code for db handler
- Move ESIResponse to a seperate module
- Move checker code from ESI.async_request to ESI.request
- Modification on market API with formatter
- Add more flexibility to InsertBuffer

This is not a completed commit. Bugs are expected with this commit and
will be fixed real quick.

* Add endpoint checker to ESIRequestChecker

Recently an endpoint for market order history was down, causing failure
in most tests. The goal is to avoid failure of tests due to closure of
some arbitrary endpoints in ESI, and also to inform user that certain
endpoints are down.

* Apply endpoint checker on unit tests

Unit tests are skipped if the requested endpoints are down. This reduces
unexpected failures in unit testing due to uncontrollable events.

* Add profile to .gitignore

* Add handlers skeleton to codebase

Syncing local changes with remote, and will convert to trunk-based
management.

---------

Co-authored-by: Hanbo <hbopublic@163.com>